convert AddFormatView to use same form as AddPartitionView
