+++ subiquity/ui/views/filesystem/__init__.py
@@ -20,8 +20,7 @@ configuration.
 
 """
 
-from .add_format import AddFormatView
-from .add_partition import AddPartitionView
+from .add_partition import AddFormatView, AddPartitionView
 from .disk_info import DiskInfoView
 from .disk_partition import DiskPartitionView
 from .filesystem import FilesystemView
